# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "使用gin+vue进行极速开发的全栈开发基础平台",
    "title": "Gin-Vue-Admin Swagger API接口文档",
    "contact": {},
    "version": "v2.5.9"
  },
  "paths": {
...
    "/base/login": {
      "post": {
        "produces": ["application/json"],
        "tags": ["Base"],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "用户名, 密码, 验证码",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回包括用户信息,token,过期时间",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/response.LoginResponse"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/user/SetSelfInfo": {
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "设置用户信息",
        "parameters": [
          {
            "description": "ID, 用户名, 昵称, 头像链接",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设置用户信息",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/admin_register": {
      "post": {
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "用户注册账号",
        "parameters": [
          {
            "description": "用户名, 昵称, 密码, 角色ID",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户注册账号,返回包括用户信息",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/response.SysUserResponse"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/changePassword": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "用户修改密码",
        "parameters": [
          {
            "description": "用户名, 原密码, 新密码",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.ChangePasswordReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户修改密码",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/deleteUser": {
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "删除用户",
        "parameters": [
          {
            "description": "用户ID",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.GetById"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除用户",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/getUserInfo": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "获取用户信息",
        "responses": {
          "200": {
            "description": "获取用户信息",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/getUserList": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "分页获取用户列表",
        "parameters": [
          {
            "description": "页码, 每页大小",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.PageInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页获取用户列表,返回包括列表,总数,页码,每页数量",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/response.PageResult"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/resetPassword": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "重置用户密码",
        "parameters": [
          {
            "description": "ID",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "重置用户密码",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/setUserAuthorities": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "设置用户权限",
        "parameters": [
          {
            "description": "用户UUID, 角色ID",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.SetUserAuthorities"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设置用户权限",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/setUserAuthority": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "更改用户权限",
        "parameters": [
          {
            "description": "用户UUID, 角色ID",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.SetUserAuth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设置用户权限",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/user/setUserInfo": {
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["SysUser"],
        "summary": "设置用户信息",
        "parameters": [
          {
            "description": "ID, 用户名, 昵称, 头像链接",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设置用户信息",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "request.ChangePasswordReq": {
      "type": "object",
      "properties": {
        "newPassword": {
          "description": "新密码",
          "type": "string"
        },
        "password": {
          "description": "密码",
          "type": "string"
        }
      }
    },
...
    "request.GetById": {
      "type": "object",
      "properties": {
        "id": {
          "description": "主键ID",
          "type": "integer"
        }
      }
    },
...
    "request.Login": {
      "type": "object",
      "properties": {
        "captcha": {
          "description": "验证码",
          "type": "string"
        },
        "captchaId": {
          "description": "验证码ID",
          "type": "string"
        },
        "password": {
          "description": "密码",
          "type": "string"
        },
        "username": {
          "description": "用户名",
          "type": "string"
        }
      }
    },
    "request.PageInfo": {
      "type": "object",
      "properties": {
        "keyword": {
          "description": "关键字",
          "type": "string"
        },
        "page": {
          "description": "页码",
          "type": "integer"
        },
        "pageSize": {
          "description": "每页大小",
          "type": "integer"
        }
      }
    },
    "request.Register": {
      "type": "object",
      "properties": {
        "authorityId": {
          "type": "string",
          "example": "int 角色id"
        },
        "authorityIds": {
          "type": "string",
          "example": "[]uint 角色id"
        },
        "email": {
          "type": "string",
          "example": "电子邮箱"
        },
        "enable": {
          "type": "string",
          "example": "int 是否启用"
        },
        "headerImg": {
          "type": "string",
          "example": "头像链接"
        },
        "nickName": {
          "type": "string",
          "example": "昵称"
        },
        "passWord": {
          "type": "string",
          "example": "密码"
        },
        "phone": {
          "type": "string",
          "example": "电话号码"
        },
        "userName": {
          "type": "string",
          "example": "用户名"
        }
      }
    },
...
    "request.SetUserAuth": {
      "type": "object",
      "properties": {
        "authorityId": {
          "description": "角色ID",
          "type": "integer"
        }
      }
    },
    "request.SetUserAuthorities": {
      "type": "object",
      "properties": {
        "authorityIds": {
          "description": "角色ID",
          "type": "array",
          "items": {
            "type": "integer"
          }
        },
        "id": {
          "type": "integer"
        }
      }
    },
...
    "response.LoginResponse": {
      "type": "object",
      "properties": {
        "expiresAt": {
          "type": "integer"
        },
        "token": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/system.SysUser"
        }
      }
    },
    "response.PageResult": {
      "type": "object",
      "properties": {
        "list": {},
        "page": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        }
      }
    },
...
    "response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "msg": {
          "type": "string"
        }
      }
    },
...
    "response.SysUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/system.SysUser"
        }
      }
    },
...
    "system.SysAuthority": {
      "type": "object",
      "properties": {
        "authorityId": {
          "description": "角色ID",
          "type": "integer"
        },
        "authorityName": {
          "description": "角色名",
          "type": "string"
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysAuthority"
          }
        },
        "createdAt": {
          "description": "创建时间",
          "type": "string"
        },
        "dataAuthorityId": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysAuthority"
          }
        },
        "defaultRouter": {
          "description": "默认菜单(默认dashboard)",
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "menus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysBaseMenu"
          }
        },
        "parentId": {
          "description": "父角色ID",
          "type": "integer"
        },
        "updatedAt": {
          "description": "更新时间",
          "type": "string"
        }
      }
    },
...
    "system.SysUser": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "主键ID",
          "type": "integer"
        },
        "activeColor": {
          "description": "活跃颜色",
          "type": "string"
        },
        "authorities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysAuthority"
          }
        },
        "authority": {
          "$ref": "#/definitions/system.SysAuthority"
        },
        "authorityId": {
          "description": "用户角色ID",
          "type": "integer"
        },
        "baseColor": {
          "description": "基础颜色",
          "type": "string"
        },
        "createdAt": {
          "description": "创建时间",
          "type": "string"
        },
        "email": {
          "description": "用户邮箱",
          "type": "string"
        },
        "enable": {
          "description": "用户是否被冻结 1正常 2冻结",
          "type": "integer"
        },
        "headerImg": {
          "description": "用户头像",
          "type": "string"
        },
        "nickName": {
          "description": "用户昵称",
          "type": "string"
        },
        "phone": {
          "description": "用户手机号",
          "type": "string"
        },
        "sideMode": {
          "description": "用户侧边主题",
          "type": "string"
        },
        "updatedAt": {
          "description": "更新时间",
          "type": "string"
        },
        "userName": {
          "description": "用户登录名",
          "type": "string"
        },
        "uuid": {
          "description": "用户UUID",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "x-token",
      "in": "header"
    }
  }
}